refactor: separate logic form LoginPageTemplate and LoginForm
